Fixed the minimax issues
The issue was that when you would make the next recursive call, it would constantly be updating the board. So when you return from that call, and try to unmake the move, it would keep all the other changes the recursive call made. To fix this, I made a copy of the board before making any changes and then reset the board to that state